+++ ydb/tests/stress/kv_volume/types.h
@@ -1,0 +1,25 @@
+#pragma once
+
+#include <util/generic/string.h>
+#include <util/system/types.h>
+
+namespace NKvVolumeStress {
+
+struct TOptions {
+    TString Endpoint = "grpc://localhost:2135";
+    TString Database;
+    ui32 Duration = 120;
+    ui32 InFlight = 1;
+    TString Version = "v1";
+    TString ConfigPath;
+    TString ConfigName;
+    bool AllowErrors = false;
+    bool Verbose = false;
+};
+
+struct TKeyInfo {
+    ui32 PartitionId = 0;
+    ui32 KeySize = 0;
+};
+
+} // namespace NKvVolumeStress
